pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kcpc"
version = "0.1.0"
description = "Keyword Competition Page Counter local batch tool"
requires-python = ">=3.10"
dependencies = ["ddgs>=6.0.0", "pandas>=2.0.0", "openpyxl>=3.1.0", "python-dotenv>=1.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
